requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlmodel
psycopg2
